+++ packages/money-account-api-data-service/jest.config.js
@@ -11,10 +11,8 @@ const baseConfig = require('../../jest.config.packages');
 const displayName = path.basename(__dirname);
 
 module.exports = merge(baseConfig, {
-  // The display name when running multiple projects
   displayName,
 
-  // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
       branches: 100,